refactor(fetcher): move ExchangeError to interceptorManager
- Relocate ExchangeError class from fetcherError.ts to interceptorManager.ts
- Remove ExchangeError import from fetcherError.ts
- Add FetcherError import to interceptorManager.ts
- Update class definition and examples in interceptorManager.ts
