+++ packages/fetcher/src/fetcherError.ts
@@ -60,44 +60,6 @@ export class FetcherError extends Error {
   }
 }
 
-/**
- * Custom error class for FetchExchange related errors.
- *
- * This error is thrown when there are issues with the HTTP exchange process,
- * such as when a request fails and no response is generated. It provides
- * comprehensive information about the failed request through the exchange object.
- *
- * @example
- * ```typescript
- * try {
- *   await fetcher.get('/api/users');
- * } catch (error) {
- *   if (error instanceof ExchangeError) {
- *     console.log('Request URL:', error.exchange.request.url);
- *     console.log('Request method:', error.exchange.request.method);
- *     if (error.exchange.error) {
- *       console.log('Underlying error:', error.exchange.error);
- *     }
- *   }
- * }
- * ```
- */
-export class ExchangeError extends FetcherError {
-  /**
-   * Creates a new ExchangeError instance.
-   *
-   * @param exchange - The FetchExchange object containing request/response/error information.
-   */
-  constructor(public readonly exchange: FetchExchange) {
-    const errorMessage =
-      exchange.error?.message ||
-      exchange.response?.statusText ||
-      `Request to ${exchange.request.url} failed during exchange`;
-    super(errorMessage, exchange.error);
-    this.name = 'ExchangeError';
-    Object.setPrototypeOf(this, ExchangeError.prototype);
-  }
-}
 
 /**
  * Custom error class for Fetcher request failures.
